feature(#255): rename modules
