--- spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-example/src/main/java/io/miragon/miranum/miranumdeploymentproxy/ExampleApplication.java
@@ -1,14 +1,0 @@
-package io.miragon.miranum.miranumdeploymentproxy;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class ExampleApplication {
-
-    public static void main(final String[] args) {
-        SpringApplication.run(ExampleApplication.class, args);
-    }
-
-}
