--- spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-example/src/test/java/io/miragon/miranum/miranumdeploymentproxy/PropertyFactory.java
@@ -1,23 +1,0 @@
-package io.miragon.miranum.miranumdeploymentproxy;
-
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
-import org.springframework.core.env.PropertySource;
-import org.springframework.core.env.PropertiesPropertySource;
-import org.springframework.core.io.support.EncodedResource;
-import org.springframework.core.io.support.PropertySourceFactory;
-
-import java.io.IOException;
-import java.util.Properties;
-
-public class PropertyFactory implements PropertySourceFactory {
-    @Override
-    public PropertySource<?> createPropertySource(String name, EncodedResource encodedResource)
-        throws IOException {
-        YamlPropertiesFactoryBean factory = new YamlPropertiesFactoryBean();
-        factory.setResources(encodedResource.getResource());
-
-        Properties properties = factory.getObject();
-
-        return new PropertiesPropertySource(encodedResource.getResource().getFilename(), properties);
-    }
-}
