--- spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-starter/pom.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <artifactId>miranum-deployment-proxy-starter</artifactId>
-    <name>miranum-deployment-proxy-starter</name>
-    <modelVersion>4.0.0</modelVersion>
-    <packaging>jar</packaging>
-
-    <!-- link the parent -->
-    <parent>
-        <groupId>io.miragon.miranum</groupId>
-        <artifactId>miranum-deployment-proxy-parent</artifactId>
-        <version>0.1.4</version>
-    </parent>
-
-    <dependencies>
-        <!-- Spring boot libraries -->
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot</artifactId>
-            <version>${spring.boot.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-autoconfigure</artifactId>
-            <version>${spring.boot.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-configuration-processor</artifactId>
-            <version>${spring.boot.version}</version>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
-            <version>${spring.boot.version}</version>
-        </dependency>
-
-        <!-- miranum-deployment-proxy library -->
-        <dependency>
-            <groupId>io.miragon.miranum</groupId>
-            <artifactId>miranum-deployment-proxy</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
-</project>
