--- spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-starter/src/main/java/io/miragon/miranum/miranumdeploymentproxy/properties/DeploymentProxyProperties.java
@@ -1,15 +1,0 @@
-package io.miragon.miranum.miranumdeploymentproxy.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import java.util.Map;
-
-@Setter
-@Getter
-@ConfigurationProperties(prefix = "io.miragon.miranum.miranum-deployment-proxy")
-public class DeploymentProxyProperties {
-    private Map<String, String> deploymentHandlers;
-    private Map<String, String> targets;
-}
