+++ spring-boot-apps/miranum-deployment/miranum-deployment-example/src/main/java/io/miragon/miranum/deploymentserver/ExampleApplication.java
@@ -1,0 +1,14 @@
+package io.miragon.miranum.deploymentserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class ExampleApplication {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(ExampleApplication.class, args);
+    }
+
+}
