+++ spring-boot-apps/miranum-deployment/miranum-deployment-example/src/main/resources/application.yaml
