+++ spring-boot-apps/miranum-deployment/miranum-deployment-example/src/main/resources/banner.txt
