+++ spring-boot-apps/miranum-deployment/miranum-deployment-example/src/test/resources/application-test.yaml
