+++ spring-boot-apps/miranum-deployment/miranum-deployment-server-core/src/main/java/io/miragon/miranum/deploymentserver/handler/DeploymentHandler.java
@@ -1,0 +1,9 @@
+package io.miragon.miranum.deploymentserver.handler;
+
+import io.miragon.miranum.deploymentserver.dto.DeploymentDto;
+import io.miragon.miranum.deploymentserver.dto.DeploymentSuccessDto;
+
+public interface DeploymentHandler {
+
+    DeploymentSuccessDto deploy(final DeploymentDto deploymentDto);
+}
