+++ spring-boot-apps/miranum-deployment/miranum-deployment-server-starter/src/main/java/io/miragon/miranum/deploymentserver/properties/DeploymentProxyProperties.java
@@ -1,0 +1,15 @@
+package io.miragon.miranum.deploymentserver.properties;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.Map;
+
+@Setter
+@Getter
+@ConfigurationProperties(prefix = "io.miragon.miranum.miranum-deployment-proxy")
+public class DeploymentProxyProperties {
+    private Map<String, String> deploymentHandlers;
+    private Map<String, String> targets;
+}
