+++ spring-boot-apps/miranum-deployment/miranum-deployment-server-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  io.miragon.miranum.deploymentserver.configuration.DeploymentProxyAutoConfiguration
